refactor: use a single ffi for gnark and circom verifiers
