--- crates/batcher/circom/go.mod
@@ -1,10 +1,0 @@
-module circom
-
-go 1.22.3
-
-require github.com/yetanotherco/go-circom-prover-verifier v0.0.0-20250618185957-f01a8a8ec4a6
-
-require (
-	github.com/ethereum/go-ethereum v1.14.0 // indirect
-	golang.org/x/sys v0.19.0 // indirect
-)
